Begone you bugs

+++ src/main/java/com/astrogreg/gregvaults/screen/AggregatedStack.java
@@ -13,10 +13,10 @@ public class AggregatedStack {
     public AggregatedStack(ItemStack displayStack, List<Integer> backingSlots) {
         this.displayStack = displayStack;
         this.backingSlots = backingSlots;
-        this.totalCount = displayStack.getCount();
+        this.totalCount   = displayStack.getCount();
     }
 
-    public void addCount(int count) {
+    public void addCount(long count) {
         this.totalCount += count;
     }
